Python README.md
I am very new to programming languages, have done CCNA and make my way to it because love games and wanna make one in Python. Also interested in C++ and C#. Unity and Unreal engine. Let's see how this platform help me in this journey. This platform is a gem.